+++ src/security/IIdentity.ts
@@ -1,0 +1,6 @@
+export interface IIdentity<T> {
+
+    id: string|number|object;
+    isAuthenticated: boolean;
+    details: T;
+}