+++ src/security/SecurityException.ts
@@ -1,0 +1,8 @@
+import { CodedException } from "../exceptions/index";
+
+export class SecurityException extends CodedException {
+    constructor(code: number, message: string, innerError?: Error) {
+        super(code, message, innerError);
+        this.name = "SecurityException";
+    }
+}